scripts/reindexing/requirements.txt
# Copyright The Linux Foundation and each contributor to LFX.
# SPDX-License-Identifier: MIT

boto3>=1.34.0
httpx>=0.27.0
nats-py>=2.7.0
opensearch-py>=2.4.0
